pyproject.toml
[build-system]
requires = ["setuptools>=64", "setuptools_scm[toml]>=8"]
build-backend = "setuptools.build_meta"

[project]
name = "scikit-learn-models"
dynamic = ["version"]
description = "Extra models for scikit-learn"
readme = "README.md"
authors = [{name="Facundo Manuel Quiroga", email="{first_name}{last_name_initial}@gmail.com"}]
requires-python = ">=3.12"
dependencies = [
    "numpy>=2.2.0",
    "matplotlib>=3.10.0",
    "graphviz>=0.20.3",
    "pandas>=2.2.3",
    "scikit-learn>=1.6.1",
    "data-science-types>=0.2.23",
    "pytest>=8.3.5",
    "pygraphviz>=1.14",
]
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "License :: OSI Approved :: BSD License",
    "Operating System :: POSIX",
    "Operating System :: Unix",
    "Operating System :: MacOS",
    "Operating System :: Microsoft :: Windows",
]

[dependency-groups]
dev = [
    "pandas-stubs>=2.2.3.250308",
    "scipy-stubs>=1.15.2.1",
    "microsoft-python-type-stubs",
    "jupyter>=1.1.1",
    "scikit-learn-models",
    "tqdm>=4.67.1",
    "types-tqdm>=4.67.0.20250401",
    "snakeviz>=2.2.2",
    "openml>=0.15.1",
    "lets-plot>=4.6.2",
    "py-cpuinfo>=9.0.0",
    "cairosvg>=2.7.1",
    "pygraphviz>=1.14",
    "tabulate>=0.9.0",
    "pandas[pyarrow]>=2.2.3",
    "pyarrow>=19.0.1",
    "black>=25.1.0",
]


[project.urls]
"Homepage" = "https://github.com/facundoq/scikit-learn-models"
"Bug Reports" = "https://github.com/facundoq/scikit-learn-models/issues"
"Source" = "https://github.com/facundoq/scikit-learn-models"

[tool.setuptools_scm]
version_file = "sklearnmodels/_version.py"

[tool.pixi.project]
channels = ["conda-forge"]
platforms = ["win-64", "linux-64", "osx-64", "osx-arm64"]


[[tool.uv.index]]
name = "pypi"
url = "https://pypi.org/simple/"
publish-url = "https://pypi.org/legacy/"

[tool.uv.sources]
microsoft-python-type-stubs = { git = "https://github.com/microsoft/python-type-stubs.git" }
scikit-learn-models = { workspace = true }


[tool.pixi.feature.lint.dependencies]
# The version below should be aligned with the one of `.pre-commit-config.yaml`
black = "23.3.0"
pre-commit = "4.3.0"
ruff = "0.4.2"

[tool.pixi.feature.lint.tasks]
black = { cmd = "black --check --diff sklearnmodels && black --check --diff examples" }
ruff = { cmd = "ruff check --output-format=full sklearnmodels && ruff check --output-format=full examples" }
lint = { depends_on = ["black", "ruff"]}


[tool.setuptools]
py-modules = ['sklearnmodels']

[tool.hatch.build.targets.wheel]
packages = ["sklearnmodels"]

[tool.pixi.feature.test.dependencies]
pytest = "*"
pytest-cov = "*"

[tool.pixi.feature.test.tasks]
test = { cmd = "pytest -vsl --cov=sklearnmodels --cov-report=xml sklearnmodels" }

[tool.pixi.feature.doc.dependencies]
matplotlib = "*"
numpydoc = "*"
pydata-sphinx-theme = "*"
setuptools-scm = ">=8"  # needed for the versioning
sphinx = "*"
sphinx-design = "*"
sphinx-gallery = "*"
sphinx-prompt = "*"

[tool.pixi.feature.doc.tasks]
build-doc = { cmd = "make html", cwd = "doc" }
clean-doc = { cmd = "rm -rf _build", cwd = "doc" }

[tool.pixi.environments]
doc = ["doc"]
lint = ["lint"]
test = ["test"]
dev = ["doc", "lint", "test"]

[tool.black]
line-length = 88
target_version = ['py311']
preview = true
exclude = '''
/(
    \.eggs         # exclude a few common directories in the
  | \.git          # root of the project
  | \.vscode
)/
'''
force-exclude = "sklearnmodels/_version.py"

[tool.ruff]
# max line length for black
line-length = 88
target-version = "py311"
exclude=[
    ".git",
    "__pycache__",
    "dist",
    "doc/_build",
    "doc/auto_examples",
    "build",
    "sklearnmodels/_version.py",
]

[tool.ruff.lint]
# all rules can be found here: https://beta.ruff.rs/docs/rules/
select = ["E", "F", "W", "I"]
ignore=[
    # space before : (needed for how black formats slicing)
    "E203",
    # do not assign a lambda expression, use a def
    "E731",
    # do not use variables named 'l', 'O', or 'I'
    "E741",
]

[tool.ruff.lint.per-file-ignores]
# It's fine not to put the import at the top of the file in the examples
# folder.
"examples/*"=["E402"]
"doc/conf.py"=["E402"]
"doc/_templates/numpydoc_docstring.py"=["F821", "W292"]

[tool.pytest.ini_options]
addopts = "--doctest-modules --color=yes"
doctest_optionflags = "NORMALIZE_WHITESPACE"

[tool.mypy]
implicit_optional = true
disallow_untyped_defs = false
check_untyped_defs = true
warn_return_any = true
